worker_node/requirements.txt
# VeriTask 3.0 — Worker Node Dependencies
# Install: pip install -r requirements.txt

dstack-sdk>=0.1.0
eth-account>=0.13.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
requests>=2.31.0
python-dotenv>=1.0.0
